[level 0] Title: 숨어있는 숫자의 덧셈 (1), Time: 0.08 ms, Memory: 87.5 MB -BaekjoonHub

+++ 프로그래머스/0/120851. 숨어있는 숫자의 덧셈 （1）/숨어있는 숫자의 덧셈 （1）.java
@@ -4,7 +4,7 @@ class Solution {
         
         for (char n : my_string.toCharArray()) {
             if(Character.isDigit(n))
-                answer += (int)n - '0';
+                answer += n - '0';
             else continue;
         }
